Move PullerProgress declaration to events.
Get rid of the now empty lib/model/types package again.
