--- lib/model/types/types.go
@@ -1,20 +1,0 @@
-// Copyright (C) 2022 The Syncthing Authors.
-//
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// You can obtain one at https://mozilla.org/MPL/2.0/.
-
-package types
-
-// A momentary state representing the progress of the puller
-type PullerProgress struct {
-	Total                   int   `json:"total"`
-	Reused                  int   `json:"reused"`
-	CopiedFromOrigin        int   `json:"copiedFromOrigin"`
-	CopiedFromOriginShifted int   `json:"copiedFromOriginShifted"`
-	CopiedFromElsewhere     int   `json:"copiedFromElsewhere"`
-	Pulled                  int   `json:"pulled"`
-	Pulling                 int   `json:"pulling"`
-	BytesDone               int64 `json:"bytesDone"`
-	BytesTotal              int64 `json:"bytesTotal"`
-}
